tweet_reader/requirements.txt
attrs
brotlipy==0.7.0
certifi
cffi
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.1
cryptography
idna==2.9
iniconfig
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe
packaging
pluggy
py
pycparser
pyOpenSSL
pyparsing
PySocks
python-dotenv
requests==2.28.1
tomli
tqdm==4.64.0
urllib3
